analysis: move driver tracking logic to AnalysisManager
